chat_rag_app_/requirements.txt
    streamlit
    langchain
    langchain-community
    langchain-core
    langchain-text-splitters
    google-generativeai
    ollama
    faiss-cpu
    pypdf # For PDF document loading
